SWBIT-4521 #comment Separates PlanOption from SignupPlanSelection
